feat: add lazy loading in SingleArtwork page
